+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Default Rest Engine supports gzip encoded content
+- Default Rest Engine supports gzip/deflate
 - Default Rest Engine adds Accept: application/json by default
 
 ### Changed